requirements.txt
flask==3.0.3
python-dotenv==1.0.1
groq>=0.11.0
gunicorn==22.0.0